fix: PHP code style violations
PHP CodeSniffer complains about
- ERROR: You must use "/**" style comments for a file comment
- ERROR: Empty line not required before block comment

Both are connected to the SPDX information achieve reuse compatibility
in aed280b35ef068201d30d526d859283783209901 as the reuse-tool adds the
license info as follows:

  <?php

  // SPDX-FileCopyrightText: YEAR AUTHOR
  //
  // SPDX-License-Identifier: LICENSE

  /**
   * FILE HEADER COMMENT
   *
  ...

Some context regarding this format can be found in
https://github.com/fsfe/reuse-tool/issues/436

Serveral ways have been considered to address CodeSniffer's errors:
a) specify SPDX information in HTML comment about the `<?php`,
b) move the SPDX information to the phpdoc file header comment, and
c) update the rulset in phpcs.xml to match the current style.

A common drawback of a) and b) is that changes to the annotation via
reuse-tool (`reuse annotate`) require manual adjustments: with a) reuse
adds additional annotations in the wrong `// comment` style; with b)
reuse-tool replaces the entire file header comment with its SPDX
comment.
Additionally, with a) the SPDX information is less coupled to the actual
PHP code, its objective.
With b) there is the question to where exactly put the SPDX information,
in the description, hijack one of the common phpdoc tags or put it in a
custom one (e.g., `@reuse`).

Eventually, we decided to go with c) which itself comes with the
drawback of having to disable the rules causing the errors, i.e.
`Squiz.Commenting.FileComment` and `Squiz.Commenting.InlineComment`.